Revert "Revert "Refactor headerRedux""
