+++ apps/src/code-studio/components/header/EditableProjectName.jsx
@@ -9,7 +9,7 @@ import {
   refreshProjectName,
   setNameFailure,
   unsetNameFailure
-} from '../../headerRedux';
+} from '../../projectRedux';
 import NameFailureDialog from '../NameFailureDialog';
 import NameFailureError from '../../NameFailureError';
 
@@ -55,7 +55,7 @@ class UnconnectedDisplayProjectName extends React.Component {
   }
 }
 const DisplayProjectName = connect(state => ({
-  projectName: state.header.projectName
+  projectName: state.project.projectName
 }))(UnconnectedDisplayProjectName);
 
 class UnconnectedEditProjectName extends React.Component {
@@ -142,8 +142,8 @@ class UnconnectedEditProjectName extends React.Component {
 }
 const EditProjectName = connect(
   state => ({
-    projectName: state.header.projectName,
-    projectNameFailure: state.header.projectNameFailure
+    projectName: state.project.projectName,
+    projectNameFailure: state.project.projectNameFailure
   }),
   {
     refreshProjectName,
