+++ apps/src/code-studio/components/header/RetryProjectSaveDialog.story.jsx
@@ -1,6 +1,6 @@
 import {UnconnectedRetryProjectSaveDialog as RetryProjectSaveDialog} from './RetryProjectSaveDialog';
 import React from 'react';
-import {projectUpdatedStatuses as statuses} from '../../headerRedux';
+import {projectUpdatedStatuses as statuses} from '../../projectRedux';
 import {action} from '@storybook/addon-actions';
 
 export default storybook => {
